Refactor Event component: extract Header and Body into separate components
Change-type: patch
